+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 
-All notable changes to the "vscode-circuitpython" extension will be documented in this file.
+All notable changes to the "vscode-circuitpython" extension will be documented
+in this file.
+
+## [0.0.4]
+
+- Refactored the serial monitor from an output channel to a terminal, allowing
+  interaction with the Circuit Python REPL
 
 ## [0.0.3]